pyproject.toml
[project]
name = "file-parser"
version = "0.1.0"
description = ""
authors = [
    {name = "Leo Mok",email = "leomok@smooothbiz.io"}
]
requires-python = ">=3.13,<3.14"
dependencies = [
    "fastapi (>=0.115.0,<1.0.0)",
    "uvicorn (>=0.30.0,<1.0.0)",
    "python-multipart (>=0.0.9,<1.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "paddlepaddle (>=3.0.0,<4.0.0)",
    "paddleocr (>=3.4.0,<4.0.0)",
    "pillow (>=10.0.0,<12.0.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "httpx (>=0.27.0,<1.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13,<3.14"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
